pyproject.toml
[project]
name = "radar"
version = "0.2.0"
description = "Personal Industry Intelligence CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "sqlmodel>=0.0.14",
    "pydantic-settings>=2.0.0",
    "google-genai>=0.1.0",
    "pgvector>=0.2.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "cloud-sql-python-connector[asyncpg]>=1.20.0",
]

[project.scripts]
radar = "radar.main:app"

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pgvector.*"
ignore_missing_imports = true
